requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.2
Pillow>=8.0.0
matplotlib>=3.3.2
tqdm>=4.50.0
tensorboard>=2.4.0
scikit-learn>=0.24.0
pyyaml>=5.4.0
requests>=2.25.0
opencv-python>=4.5.0 